+++ app/build.gradle
@@ -29,6 +29,7 @@ android {
     buildFeatures {
         viewBinding true
     }
+    namespace 'ru.yandex.practicum.contacts'
 }
 
 dependencies {
