+++ app/src/main/java/ru/yandex/practicum/contacts/presentation/filter/FilterContactTypeViewModel.java
@@ -96,8 +96,6 @@ public class FilterContactTypeViewModel extends BaseBottomSheetViewModel {
         }
     }
 
-    public void onFilterTypeItemClick(FilterContactTypeAdapter.FilterContactTypeUi filterContactTypeUi) {
-    }
 
     static class UiState {
         public boolean isApplyEnable = false;
