+++ app/src/main/java/ru/yandex/practicum/contacts/presentation/main/ContactAdapter.java
@@ -12,7 +12,6 @@ import androidx.core.content.ContextCompat;
 import androidx.recyclerview.widget.AdapterListUpdateCallback;
 import androidx.recyclerview.widget.AsyncDifferConfig;
 import androidx.recyclerview.widget.AsyncListDiffer;
-import androidx.recyclerview.widget.DiffUtil;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
@@ -22,9 +21,7 @@ import java.util.Objects;
 
 import ru.yandex.practicum.contacts.R;
 import ru.yandex.practicum.contacts.databinding.ItemContactBinding;
-import ru.yandex.practicum.contacts.model.ContactType;
 import ru.yandex.practicum.contacts.presentation.base.BaseListDiffCallback;
-import ru.yandex.practicum.contacts.presentation.base.ListDiffInterface;
 
 public class ContactAdapter extends RecyclerView.Adapter<ContactAdapter.ViewHolder> {
 
@@ -51,13 +48,8 @@ public class ContactAdapter extends RecyclerView.Adapter<ContactAdapter.ViewHold
         return differ.getCurrentList().size();
     }
 
-    public void setItems(List<ContactUi> items) {
-        differ.submitList(items);
-    }
-
-
-
-    public void setItems(List<ru.yandex.practicum.contacts.presentation.main.ContactUi> contacts, Runnable runnable) {
+    public void setItems(List<ContactUi> items, @NonNull Runnable callback) {
+        differ.submitList(items, callback);
     }
 
     static class ViewHolder extends RecyclerView.ViewHolder {
@@ -95,26 +87,6 @@ public class ContactAdapter extends RecyclerView.Adapter<ContactAdapter.ViewHold
                     .into(binding.contactPhoto);
         }
     }
+}
 
-        public class ContactUi implements ListDiffInterface<ContactUi> {
 
-            @Override
-                public boolean theSameAs(ContactUi contactUi) {
-                    return this.hashCode() == contactUi.hashCode();
-                }
-                    public int getName() {
-                        return 0;
-                    }
-                        public CharSequence getPhone() {
-                            return null;
-                        }
-
-                        public List<ContactType> getTypes() {
-                            return null;
-                        }
-
-                        public byte[] getPhoto() {
-                            return new byte[0];
-                        }
-                    }
-                }
