+++ app/src/main/java/ru/yandex/practicum/contacts/presentation/main/ContactUi.java
@@ -7,13 +7,14 @@ import java.util.List;
 import ru.yandex.practicum.contacts.model.ContactType;
 import ru.yandex.practicum.contacts.presentation.base.ListDiffInterface;
 
-public class ContactUi {
+
+
+    public class ContactUi implements ListDiffInterface<ContactUi> {
 
         private final String name;
         private final String phone;
         private final String photo;
         private final List<ContactType> types;
-
         public ContactUi(
                 @NonNull String name,
                 @NonNull String phone,
@@ -25,36 +26,28 @@ public class ContactUi {
             this.photo = photo;
             this.types = types;
         }
-
         public String getName() {
             return name;
         }
-
         public String getPhone() {
             return phone;
         }
-
         public String getPhoto() {
             return photo;
         }
-
         public List<ContactType> getTypes() {
             return types;
         }
-
         @Override
         public boolean equals(Object o) {
             if (this == o) return true;
             if (o == null || getClass() != o.getClass()) return false;
-
             ContactUi contact = (ContactUi) o;
-
             if (!name.equals(contact.name)) return false;
             if (!phone.equals(contact.phone)) return false;
             if (!photo.equals(contact.photo)) return false;
             return types.equals(contact.types);
         }
-
         @Override
         public int hashCode() {
             int result = name.hashCode();
@@ -65,5 +58,8 @@ public class ContactUi {
         }
 
 
+        @Override
+        public boolean theSameAs(ContactUi newItem) {
+            return this.hashCode() == newItem.hashCode();
         }
-
+    }
