+++ app/src/main/java/ru/yandex/practicum/contacts/presentation/sort/SortTypeUI.java
@@ -2,9 +2,13 @@ package ru.yandex.practicum.contacts.presentation.sort;
 
 import androidx.annotation.NonNull;
 
+import ru.yandex.practicum.contacts.presentation.base.ListDiffInterface;
 import ru.yandex.practicum.contacts.presentation.sort.model.SortType;
 
-public class SortTypeUI {
+
+    public class SortTypeUI implements ListDiffInterface<SortTypeUI> {
+
+
 
     private final SortType sortType;
     private final boolean selected;
@@ -39,4 +43,9 @@ public class SortTypeUI {
         result = 31 * result + (selected ? 1 : 0);
         return result;
     }
-}
+        @Override
+        public boolean theSameAs(SortTypeUI sortTypeUI) {
+            return this.getSortType() == sortTypeUI.getSortType();
+        }
+    }
+
